refactor(connector/debezium): mysql metadata retrieval
* Changing the design of datashape retrieval adding DebeziumDatashapeStrategy interface. It will be easier to include any different strategy for retrieval (based on the target database, for instance)
* Moving old metadata retrieval logic used in POC (targeted for mysql) into DebeziumMySQLDatashapeStrategy

Ref ENTESB-13752
